+++ tests/integration/css-modules/fixtures/global-module/src/App.jsx
@@ -1,0 +1,8 @@
+/* eslint-disable node/no-unsupported-features/es-syntax */
+import React from 'react';
+import styles from './index.module.css';
+
+const App = () => <div id="verify-div" className={styles.foo} />;
+
+export default App;
+/* eslint-enable node/no-unsupported-features/es-syntax */