+++ tests/integration/css/fixtures/exclude-less/modern.config.js
@@ -1,0 +1,10 @@
+/* eslint-disable import/no-commonjs */
+module.exports = {
+  tools: {
+    less: (opts, { addExcludes }) => {
+      addExcludes([/b\.less$/]);
+    },
+  },
+};
+
+/* eslint-enable import/no-commonjs */