+++ tests/integration/css/fixtures/less-import/src/App.jsx
@@ -1,0 +1,3 @@
+import './app.less';
+
+export default () => 'less-import';