+++ tests/integration/css/fixtures/multi-css/src/entry1/App.jsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import './a.css';
+
+const App = () => <div id="entry1">entry1 app</div>;
+
+export default App;