+++ packages/generator/generators/eslint-generator/CHANGELOG.md
@@ -1,6 +1,13 @@
 # @modern-js/eslint-generator
 
+## 1.1.2
+
+### Patch Changes
+
+- 2c0750e1: fix generator export config
+
 ## 1.1.1
+
 ### Patch Changes
 
 - 687c92c7: refactor: generator input questions